requirements.txt
# K-JULIA AI Voice Automation System
# Core Dependencies

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
supabase==2.3.0
python-dotenv==1.0.0

# Voice Processing (optional - requires audio hardware)
# SpeechRecognition==3.10.0
# PyAudio==0.2.14
# pyttsx3==2.90

# Utilities
requests==2.31.0

# WSGI Server for production
gunicorn==21.2.0
